Removed needless {{ .Site.BaseUrl }}
It is not the problem if baseurl(of config.toml) ends with domain like 'http://blog.com/'.
But it breaks theme if you give it a path like 'http://blog.com/casper'.

For example:
If baseurl in config is 'http://blog.com/casper'
the code below in layouts/partials/header.html
    <link rel="stylesheet" type="text/css" href="{{.Site.BaseUrl}}css/screen.css" />

will be
    <link rel="stylesheet" type="text/css" href="http://blog.com/caspercss/screen.css" />

Even if  you change the baseurl to 'http://blog.com/casper/' it does not work at all.
Just the same thing happens.

so I've removed needless part and the first code has become
    <link rel="stylesheet" type="text/css" href="css/screen.css" />

It works for all of 'http://blog.com/' and 'http://blog.com/casper'.