Support excerpt variable
Hugo generates summaries for posts which can be used with the .Summary
page variable. The theme uses this to print the summary of each post
in the post list. As it is now, you cannot disable the summary
from being printed, so the partial header li.html is changed
to support the excerpt variable in posts.

When posts do not have an excerpt variable, the behavior is not changed
and the summary is being printed as before. If excerpt exists
and is not empty, then it is printed instead of the summary and excerpt
is empty, then nothing is printed.