Fix multiple author on front page
The logic for rendering authors on the front page had issues. The authors set on the post frontmatter were not showing(ignored) as the template logic was trying to override a `$author` variable in a subtle manner.

This fixes the issue and ensures the right author is rendered i.e when the author is set on the frontmatter then he will be picked, and if no author is specified it will default to the site wide author defined in `.Site.Params.author`.